feat: Make code blocks focusable
While generating the code, it can exceed the screen width. This forces
browser to show horizontal scroll. In fact, user can not interact with
the code block as it is not marked as focusable. This raises the accessibility
issue when users can not focus and scroll the code block with keyboard.

(see more https://dequeuniversity.com/rules/axe/4.8/scrollable-region-focusable?application=AxeChrome)

Adding tabIndex will resolve this issue and make the code block focusable